+++ src/filters/BaseElementFilter.php
@@ -16,7 +16,7 @@ use yii\base\NotSupportedException;
  * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
  * @since 2.1.0
  */
-abstract class BaseElementFilter extends Component implements FilterInterface
+abstract class BaseElementFilter extends Component implements ExclusiveFilterInterface
 {
     public static function show(string $class, string $event): bool
     {
@@ -33,6 +33,11 @@ abstract class BaseElementFilter extends Component implements FilterInterface
                 Elements::EVENT_AFTER_UPDATE_SLUG_AND_URI,
             ])
         ));
+    }
+
+    public static function excludes(): array
+    {
+        return [];
     }
 
     public static function check(Event $event, bool $value): bool